refactored the slot iterators and moved those to rocks
